# Add benchmarks and improve performance