Hi!

**Changelog**

This PR adds benchmarks comparing `json.Marshal` and `structhash.Dump`.

There are three kinds of benchmarks:
* simple: serialize struct without struct tags, maps, slices, and nested structs
* full: serialize struct without struct tags, but with maps, slices, and nested structs
* tags: serialize struct with struct tags

This PR also introduces some obvious optimizations:
* use `bytes.Buffer` instead of string concatenation
* use more compact encoding (and drop indentation)
* try to reduce number of allocations

**Results**

*Before this PR*

```
BenchmarkSimpleJSON-4	  500000	      3066 ns/op
BenchmarkSimpleDump-4	  100000	     11745 ns/op
BenchmarkFullJSON-4  	  100000	     16801 ns/op
BenchmarkFullDump-4  	   20000	     85045 ns/op
BenchmarkTagsJSON-4  	 1000000	      1763 ns/op
BenchmarkTagsDump-4  	  200000	     10534 ns/op
```

*After this PR*

```
BenchmarkSimpleJSON-4	  500000	      3039 ns/op
BenchmarkSimpleDump-4	  300000	      4910 ns/op
BenchmarkFullJSON-4  	  100000	     16665 ns/op
BenchmarkFullDump-4  	   50000	     29468 ns/op
BenchmarkTagsJSON-4  	 1000000	      1761 ns/op
BenchmarkTagsDump-4  	  300000	      5390 ns/op
```

**Further improvements**

This PR improves performance (~2x ans sometimes more), but `structhash` is still slower than `json.Marshal`, which is not the [fastest serializer](https://github.com/alecthomas/go_serialization_benchmarks) itself.

Why `structhash` is slower:
* `reflect` is slow, and `json.Marshal` caches information about types, while `structhash` doesn't
* `structhash` probably performs some unnecessary allocations
* `structhash` sorts struct keys
* `structhash` handles struct tags inefficiently

To improve performance, we can:
* cache type information retrieved from reflect
* cache sorted struct fields
* improve encoding even more (probably switch to binary encoding)
* improve struct tags parsing
* try to reduce allocations even more

I believe it's possible to make `structhash` as fast as other reflection-based serializers (e.g. [this one](https://github.com/vmihailenco/msgpack)) which are several times faster than `json.Marshal`.