@gavv 2016-05-29
Note: you could also add `go test -bench .` to travis build to check that benchmarks don't fail.