@gavv 2016-06-08
*NB*

Actually, it should be much more efficient to update resulting hash incrementally instead of serializing struct and then calculating its hash.

I think this approach, combined with reflection caching, could be much faster than `json.Marshal` and other serializers.

I found another module, similar to `structhash`, that implements incremental hashing: [`mitchellh/hashstructure`](https://github.com/mitchellh/hashstructure). (but it doesn't implement caching too)