@cnf 2016-06-09
yeah, http://github.com/mitchellh tends to write nice code! 

I might look into it when I got some free time. Thanks for the contributions so far!