@mitchellh 2016-06-09
Let me know @cnf, @gavv if you have any feedback to improve performance of hashstructure. I'd be happy to do so.