@cnf 2016-06-09
ha!, hi @mitchellh! and if you have any feedback on structhash do let me know :P Or maybe it would make more sense to merge the features?