@mitchellh 2016-06-09
Up to you, structhash didn't exist when I needed hashstructure (I think a few months ago?). I don't mind alternate implementations, but if you want to talk about merging the functionality I'm open to it and seeing how it fits.