@cnf 2016-06-09
https://github.com/cnf/structhash/commit/144eaa644e86d7b89f9309dd427d5f5cb0fa830f initial commit was Dec 9, 2014.

ATM I am not using it myself, though, so i'm thinking your tree will see better maintenance. Basically, @gavv is the activity on structhash atm :P