@gavv 2016-06-09
@mitchellh, I didn't run any benchmarks on `hashstructure` yet, but I expect it can benefit from using caching for reflection (as `json.Marshal` does it) and probably from using [`xxhash`](https://github.com/OneOfOne/xxhash).

BTW, I was unable to find both `structhash` and `hashstructure` using google (my google fu isn't perfect though). I've found both of them by guessing the name and checking it in [golanglibs.com](https://golanglibs.com/).