@cnf 2016-06-09
And thank you for showing me [golanglibs.com](https://golanglibs.com) @gavv :) I tend to try [godocs.com](https://godocs.com) as well for searching.